+++ test/unit/test_timestamp.gd
@@ -1,7 +1,7 @@
 extends GutTest
 
 func test_utc_timestamp_includes_milliseconds_and_z_suffix() -> void:
-	var client = load("res://addons/auralog/auralog.gd").new()
+	var client = load("res://addons/auralogs/auralogs.gd").new()
 	add_child_autofree(client)
 	var timestamp: String = client.call("_utc_timestamp")